opencv-video-undistorter/requirements.txt
altair==4.2.0
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
pathlib==1.0.1
pi==0.1.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.0
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
pyheif==0.7.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.0.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
streamlit-ext==0.1.5
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
Werkzeug==2.2.2
zipp==3.11.0
